Move multicore stuff to the `multicore` module
